requirements.txt
Flask==3.0.3
pytesseract==0.3.8
Pillow==10.3.0
langdetect==1.0.9
